% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sendsay}
\alias{sendsay}
\alias{sendsay-package}
\alias{sendsay}
\title{Sendsay API}
\usage{
sendsay(..., .session = NULL, .host = .SENDSAY_DEFAULT_HOST)
}
\arguments{
\item{...}{API request parameters}

\item{.session}{\itemize{
\item session
}}

\item{.host}{\itemize{
\item host
}}
}
\value{
response for the API request
}
\description{
Minimal wrapper to access Sendsay's API.

This is an extremely minimal client. You need to know the API
to be able to use this client.
}
\details{
to to to
}
\examples{
sendsay(action="ping")
sendsay(action="login", login="demo", passwd="demo")
sendsay(action="login", login="demo", passwd="demo", .host="https://api.sendsay.ru")
sendsay(action="ping", .host="https://api.sendsay.ru")
sendsay(action="pong", .host="https://api.sendsay.ru", .session="6d47ca72-e09c-42b5-b2bb-2dc88843adb1")

}
